styling: refactored interlacer utility functions into interlacer_utils.py
