+++ tests/integration/dicom/test_interlacer.py
@@ -7,6 +7,8 @@ from pathlib import Path
 from imgtools.dicom import Interlacer
 from imgtools.dicom.crawl import Crawler
 
+from imgtools.utils.interlacer_utils import SeriesNode, visualize_forest, print_interlacer_tree
+
 @pytest.mark.skipif(
     os.getenv("TEST_DATASET_TYPE", "public").lower() == 'public',
     reason="Test uses collections in private data"
@@ -96,7 +98,7 @@ def test_interlacer_visualize(medimage_by_collection, caplog) -> None:
         interlacer = Interlacer(crawler.index)
 
         viz_path = Path(series_list[0].get('Path')).parent.parent / collection / "interlacer.html"
-        interlacer.visualize_forest(viz_path)
+        visualize_forest(interlacer.root_nodes, viz_path)
 
         assert viz_path.exists()
